requirements.txt
streamlit==1.27.0
openai==0.28.1
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.2
pandas==2.0.3
